requirements.txt
python-telegram-bot==20.4
pocketoptionapi-async==1.0.1

requests
pocketoptionapi-async
